+++ extensions/ql-vscode/src/common/commands.ts
@@ -32,6 +32,13 @@ export type SingleSelectionCommandFunction<Item> = (
  * the implementation in the corresponding `getCommands` function.
  */
 
+// Builtin commands where the implementation is provided by VS Code and not by this extension.
+// See https://code.visualstudio.com/api/references/commands
+export type BuiltInVsCodeCommands = {
+  "markdown.showPreviewToSide": (uri: Uri) => Promise<void>;
+  "workbench.action.reloadWindow": () => Promise<void>;
+};
+
 // Base commands not tied directly to a module like e.g. variant analysis.
 export type BaseCommands = {
   "codeQL.openDocumentation": () => Promise<void>;
@@ -233,7 +240,8 @@ export type MockGitHubApiServerCommands = {
   "codeQL.mockGitHubApiServer.unloadScenario": () => Promise<void>;
 };
 
-export type AllCommands = BaseCommands &
+// All commands where the implementation is provided by this extension.
+export type AllExtensionCommands = BaseCommands &
   QueryEditorCommands &
   ResultsViewCommands &
   QueryHistoryCommands &
@@ -247,6 +255,8 @@ export type AllCommands = BaseCommands &
   SummaryLanguageSupportCommands &
   Partial<TestUICommands> &
   MockGitHubApiServerCommands;
+
+export type AllCommands = AllExtensionCommands & BuiltInVsCodeCommands;
 
 export type AppCommandManager = CommandManager<AllCommands>;