+++ extensions/ql-vscode/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## [UNRELEASED]
 
-- Show data flow paths of a variant analysis in a new tab
-- Show labels of entities in exported CSV results [#2170](https://github.com/github/vscode-codeql/pull/2170)
+- Show data flow paths of a variant analysis in a new tab. [#2172](https://github.com/github/vscode-codeql/pull/2172) & [#2182](https://github.com/github/vscode-codeql/pull/2182)
+- Show labels of entities in exported CSV results. [#2170](https://github.com/github/vscode-codeql/pull/2170)
 
 ## 1.8.0 - 9 March 2023